"Pass client object instead of clientId to create menu button"
Adjusted the ClientElementsFactory to pass the full client object instead of just the clientId when creating the MenuFicheClientButton. This was changed to allow the FicheClientView to take in the client object, enabling the client's full name to be displayed on the button text. This provides a more user-friendly interface by showing the client's name instead of their ID.